// src/app/core/models/BusinessSubscription/businessSub.model.ts
export class BusinessSubModel {
    _id: string;
    company_name: string;
    head_office: string;
    siren: string;
    email: string;
    creation_date: Date;
    address: string;
    vat: any;
    company_phone: number;
    first_name: string;
    birthdate: Date;
    last_name: string;
    password: string;
    subscription_id: string;
    subscription_plan_id: string;
    user_name: string;
    conform_password: string;
    reference_source: string;
    company_code: string;
    numberOfEmployee: number;
    totalempamount: number;
}